Handle particular case with IF and statement assignments.
IF that don't contains any assignments but depends on a variable used in the SCA process won't be wrapped and handled as an exception. This is usually used to `ALLOCATE` memory.
In the case a vector is assigned to another directly (without the use of indexes), we shouldn't wrap it in a loop. `a(1:10); b(1:10); a = b`