Add Fusion-Scalar transformation to SCA (on branch) (#427)
* Add Fusion-Scalar transformation.

This method group adjacent loop together and doesn't promote variable if possible.

* Xmod was changed to Module.

* Handle IF statement separatelly.

* Use List of Set since Array of Set is discouraged.

* Fix formatting and PartialCPU method description.

* Fixed redundant access to IF statement.
Renamed CPUAnalysis to CPUExtraction, since it define better its use.

* Updated references to match the current expected results.

* Handle particular case with IF and statement assignments.

IF that don't contains any assignments but depends on a variable used in the SCA process won't be wrapped and handled as an exception. This is usually used to `ALLOCATE` memory.
In the case a vector is assigned to another directly (without the use of indexes), we shouldn't wrap it in a loop. `a(1:10); b(1:10); a = b`

* Cleanup and rename of utility method.

* Remove previous CPU transformation method from source.

We don't use it ATM so we don't need to keep it in the class.

* Fix a regression introduce to handle a special case.

In order to not add an unnecessary DO statement we didn't add a necessary one.

* Handle both single and fusion transformation via config file.

Revert "Remove previous CPU transformation method from source."
This reverts commit 4ecfc407e3b2137f2fadc5fc5d8e0bbdabf89d13.

* Fix Javadoc and misleading comment.

* Better handle IF groups detection.

* Fix incorrect end of DO block.

If a statement which should wrapped in a DO is followed by a IF containing any promoted statements, the IF would be wrapped together with the statement before it.
This is incorrect if the statement inside the IF is already inside a DO and should be handled separately.

* Fix promotion condition.

We intersected region vars with the intersection vars for that region, which is useless.

* Remove transformation counting.

It also block execution and promotion if there is only one region, so when going from 1D to ?D.